Added the skeleton and the generic code/logic for handling the dpif provider functions.
Added the code in order of the functions declarations/definitions in the dpif-provider class.

Last methods to add the generic and skeleton code uf operate.

Also put suggestions on how to implement the actually logic on xdp in the comments, but this is subject to change as we implement more of the methods and get a better context.