Added code to send an upcall packet to the datapath using netdev send.
For the xdp flow struct added logic to pad the length of the structs to be multiples of u64. The complier pads these structs but when the size if calculated or used, the number padded size is used. This was resulting in some inconsistency in values read from pointers. Also in some cases a program load failure. A comment in the code explains more.

Also added function to log the keys for debug purposes.

Added necessary helpers and refactored the automake.mk