Changed the xdp code to use btf for portability and ease of definition of maps.
Adapted the loader and the header libraries and functions from the xdp-tools git repo. The repo included a library libxdp hence had to change the name of the library generated by our code. Since the proposed name for now is xdp switch (xswitch) changed all the respective xdp/ to xswitch/ and the libxdp to libxswitch and xdp flow (xf)

The adapted loader was getting an error for the program when trying to load an xdp program that writes to the packet. The error "cannot write into packet" was resolved with a patch similar to the one in link (https://lkml.org/lkml/2020/8/18/1261) and is available from linux 5.9, so the current build depends on it.

Removed dependencies on <linux/if_link.h> due to conflict of nla_bitfield32.

Yet to link the loader and the xf definitions with the dpif-xdp in OVS ./lib