Migrated to btf maps and using the xdp dispatcher to load programs.
Letting ARP be handled by userspace program.

The output now uses bpf_redirect_map but it's depended on an updated ARP table. Currently the userspace way of handling ARP does not update the ARP table.

An option is to add a bridge and do XDP_PASS for the arp but the OVS integration doesn't seem to create a bridge but rather a tap device. SHould look into that.

We can also update the ARP table when we create the namespaces but this may only be for testing.

Using perf to send logs to userspace from the xdp driver.

Creating multiple memory for xsk sockets, should look into shared memory area.