Increased the number of interfaces for tx_port.
Added an action _xdp_normal which will use the arp table for handling packets. However this actions is not working for return packets eg icmp reply. Some debuging needed.

Solved the issue of high CPU usage due to poll_loop by adding poll_timer_wait instead of poll_immediate_wake. This is because sometimes esp when there are pending events poll() is not called hence this was resulting in 0ms poll_loop which was causing the high CPU usage.

The systems seems to be losing the initial packets of each of the sessions, icmp, udp.

Currently tcp is not working, need to debug this. This might be because of missing flow rules to TCP session management packets. Might have to handle this in userspace like for the case of arp.